requirements.txt
torch>=1.2.0
torchvision
dominate>=2.3.1
dill
scikit-image
numpy>=1.15.4
scipy>=1.1.0
matplotlib
opencv-python>=3.4.3.18
tensorboard==1.14.0
tqdm
librosa
ffmpeg-python
lws
face-alignment
